requirements.txt
boto==2.49.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.3
django-extensions==2.1.4
django-widget-tweaks==1.4.3
google-cloud-storage==1.13.0
gunicorn==19.9.0
ipython==7.2.0
psycopg2-binary==2.7.6.1
python-decouple==3.1
pytz
pylint==2.2.2
